# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Global Diplomatic Relations platform API
servers:
- url: /api
  description: Base API path
...
paths:
...
  /countries:
    get:
      operationId: listCountries
      tags:
      - countries
      summary: List countries
      parameters:
      - $ref: '#/components/parameters/Search'
      - name: region
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
    post:
      operationId: createCountry
      tags:
      - countries
      summary: Create a country workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
  /countries/{id}:
    get:
      operationId: getCountry
      tags:
      - countries
      summary: Get a country workspace
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '404':
          description: Country not found
    patch:
      operationId: updateCountry
      tags:
      - countries
      summary: Update a country workspace
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryUpdate'
      responses:
        '200':
          description: Updated country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Country'
        '404':
          description: Country not found
  /contacts:
    get:
      operationId: listContacts
      tags:
      - contacts
      summary: List diplomatic contacts
      parameters:
      - $ref: '#/components/parameters/Search'
      - name: countryId
        in: query
        schema:
          type: integer
      - name: status
        in: query
        schema:
          type: string
          enum:
          - verified
          - review
          - outdated
      responses:
        '200':
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      operationId: createContact
      tags:
      - contacts
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
  /meetings:
    get:
      operationId: listMeetings
      tags:
      - meetings
      summary: List meetings
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - scheduled
          - completed
          - follow_up
      - name: countryId
        in: query
        schema:
          type: integer
      responses:
        '200':
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Meeting'
    post:
      operationId: createMeeting
      tags:
      - meetings
      summary: Schedule a meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
  /meetings/{id}:
    patch:
      operationId: updateMeeting
      tags:
      - meetings
      summary: Update a meeting
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingUpdate'
      responses:
        '200':
          description: Updated meeting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
        '404':
          description: Meeting not found
  /agreements:
    get:
      operationId: listAgreements
      tags:
      - agreements
      summary: List agreements
      parameters:
      - $ref: '#/components/parameters/Search'
      - name: status
        in: query
        schema:
          type: string
          enum:
          - draft
          - review
          - signed
          - archived
      - name: countryId
        in: query
        schema:
          type: integer
      responses:
        '200':
          description: Agreements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agreement'
    post:
      operationId: createAgreement
      tags:
      - agreements
      summary: Create an agreement record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgreementInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agreement'
  /agreements/{id}:
    patch:
      operationId: updateAgreement
      tags:
      - agreements
      summary: Update an agreement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgreementUpdate'
      responses:
        '200':
          description: Updated agreement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agreement'
        '404':
          description: Agreement not found
  /activity:
    get:
      operationId: listActivity
      tags:
      - activity
      summary: List recent activity
      parameters:
      - name: countryId
        in: query
        schema:
          type: integer
      responses:
        '200':
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
...
components:
  parameters:
    Search:
      name: search
      in: query
      schema:
        type: string
...
  schemas:
...
    Country:
      type: object
      required:
      - id
      - name
      - code
      - region
      - status
      - contactsCount
      - meetingsCount
      - riskLevel
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        region:
          type: string
        status:
          type: string
          enum:
          - active
          - leads
          - inactive
          - agreement
          - scheduled
        contactsCount:
          type: integer
        meetingsCount:
          type: integer
        riskLevel:
          type: string
          enum:
          - low
          - medium
          - high
        language:
          type:
          - string
          - 'null'
        governmentType:
          type:
          - string
          - 'null'
          enum:
          - presidential republic
          - semi-presidential
          - parliamentary republic
          - parliamentary monarchy
          - constitutional monarchy
          - absolute monarchy
          - one-party state
          - transitional
        electionYear:
          type:
          - integer
          - 'null'
        team:
          type:
          - string
          - 'null'
        priority:
          type:
          - string
          - 'null'
          enum:
          - low
          - medium
          - high
        strategy:
          type:
          - string
          - 'null'
    CountryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 2
          maxLength: 3
        region:
          type: string
          minLength: 1
        status:
          type: string
          enum:
          - active
          - leads
          - inactive
          - agreement
          - scheduled
        riskLevel:
          type: string
          enum:
          - low
          - medium
          - high
        language:
          type: string
        governmentType:
          type: string
          enum:
          - presidential republic
          - semi-presidential
          - parliamentary republic
          - parliamentary monarchy
          - constitutional monarchy
          - absolute monarchy
          - one-party state
          - transitional
        electionYear:
          type: integer
        team:
          type: string
        priority:
          type: string
          enum:
          - low
          - medium
          - high
        strategy:
          type: string
    CountryInput:
      type: object
      required:
      - name
      - code
      - region
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 2
          maxLength: 3
        region:
          type: string
          minLength: 1
        status:
          type: string
          enum:
          - active
          - leads
          - inactive
          - agreement
          - scheduled
    Contact:
      type: object
      required:
      - id
      - name
      - title
      - institution
      - countryId
      - countryName
      - email
      - verificationStatus
      - lastVerified
      - relationship
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        institution:
          type: string
        countryId:
          type: integer
        countryName:
          type: string
        email:
          type: string
        phone:
          type:
          - string
          - 'null'
        verificationStatus:
          type: string
          enum:
          - verified
          - review
          - outdated
        lastVerified:
          type: string
          format: date
        relationship:
          type: string
    ContactInput:
      type: object
      required:
      - name
      - title
      - institution
      - countryId
      - email
      - relationship
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        institution:
          type: string
          minLength: 1
        countryId:
          type: integer
        email:
          type: string
          format: email
        phone:
          type: string
        relationship:
          type: string
    Meeting:
      type: object
      required:
      - id
      - title
      - countryName
      - date
      - status
      - participants
      - actionArea
      properties:
        id:
          type: integer
        title:
          type: string
        countryName:
          type: string
        date:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - scheduled
          - completed
          - follow_up
        participants:
          type: integer
        actionArea:
          type: string
        owner:
          type: string
    MeetingInput:
      type: object
      required:
      - title
      - countryId
      - date
      - actionArea
      properties:
        title:
          type: string
          minLength: 1
        countryId:
          type: integer
        date:
          type: string
          format: date-time
        actionArea:
          type: string
          minLength: 1
        owner:
          type: string
    MeetingUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        date:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - scheduled
          - completed
          - follow_up
        actionArea:
          type: string
          minLength: 1
        owner:
          type: string
    Agreement:
      type: object
      required:
      - id
      - name
      - type
      - countryName
      - status
      - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        countryName:
          type: string
        status:
          type: string
          enum:
          - draft
          - review
          - signed
          - archived
        updatedAt:
          type: string
          format: date
        renewalDate:
          type:
          - string
          - 'null'
          format: date
    AgreementInput:
      type: object
      required:
      - name
      - type
      - countryId
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          minLength: 1
        countryId:
          type: integer
        status:
          type: string
          enum:
          - draft
          - review
          - signed
          - archived
        renewalDate:
          type: string
          format: date
    AgreementUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          minLength: 1
        status:
          type: string
          enum:
          - draft
          - review
          - signed
          - archived
        renewalDate:
          type:
          - string
          - 'null'
          format: date
    Activity:
      type: object
      required:
      - id
      - kind
      - title
      - description
      - occurredAt
      properties:
        id:
          type: integer
        kind:
          type: string
        title:
          type: string
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
        countryName:
          type:
          - string
          - 'null'
        actorId:
          type:
          - string
          - 'null'
        actorName:
          type:
          - string
          - 'null'
...